test: complete shared test setup migration
Migrate the remaining shared setup from jest.setup.js into
test/setupTests.js.

Restore the global ErrorHandler and the complete canvas mock, then
remove redundant local ErrorHandler mocks and delete the obsolete
jest.setup.js file.
