+++ js/__tests__/IntervalsBlocks.test.js
@@ -113,11 +113,6 @@ global.Singer = {
     }
 };
 
-global.ErrorHandler = {
-    warn: jest.fn(),
-    error: jest.fn()
-};
-
 const { setupIntervalsBlocks } = require("../blocks/IntervalsBlocks.js");
 
 function makeTurSinger(overrides = {}) {
