+++ js/blocks/__tests__/IntervalsBlocks.test.js
@@ -22,11 +22,6 @@
 
 const { setupIntervalsBlocks } = require("../IntervalsBlocks");
 
-global.ErrorHandler = {
-    warn: jest.fn(),
-    error: jest.fn()
-};
-
 describe("setupIntervalsBlocks", () => {
     let activity;
     let logo;
